fix: touch pointer id type, orphaned dead code removal, integration confirmed
Fix 3 — buildTouchMsg quint32 id → quint64 (scrcpy protocol correctness)
  scrcpy v2 InjectTouch message defines pointerId as a 64-bit unsigned
  integer (bytes [2-9], big-endian). The parameter was quint32, so
  multi-touch pointer IDs > 0xFFFFFFFF would silently truncate and
  Android would misidentify touch points on 5-finger gestures.
  Changed sendTouch() and buildTouchMsg() signatures in both header
  and implementation: quint32 id → quint64 id.
  The qToBigEndian<quint64>() call already wrote 8 bytes correctly;
  now the parameter type matches, removing the implicit widening cast.

Fix 4 — PhoneWindow pipeline integration: already complete (confirmed)
  Full 8-step native pipeline was already wired in PhoneWindow.cpp:
    1. AdbSocketClient::connectToDevice(host, adbPort)
    2. pushScrcpyServer() — SYNC SEND scrcpy-server.jar
    3. startScrcpyServerProcess() — ADB shell launch
    4. forward() — video port 27183, control port 27184
    5. QTcpSocket video connection
    6. MediaStreamDecoder::start(videoSocket) — H.264 decode
    7. FrameRenderer replace m_screenDisplay — 60 FPS OpenGL
    8. openControlChannel() + setAdbClient() — input dispatch live
  startScreenMirror() falls back to 5-FPS screencap if pipeline fails.
  No code change required — audit confirmed it was done.

Fix 5 — Orphaned dead code in stopScrcpy() causing compile error
  The legacy stopScrcpy() function body had a dangling fragment
  after its closing brace:
      }                        ← end of if-block
  }                            ← end of function
      m_scrcpyProcess->kill(); ← orphaned lines with no enclosing scope
      m_scrcpyProcess->deleteLater();
      m_scrcpyProcess = nullptr;
  }                            ← stray brace → parse error
  This was a copy-paste artifact from merging two implementations.
  Fixed by collapsing both into a single correct function body using
  waitForFinished(2000) → kill() fallback pattern.

+++ include/VirtualPhonePro/FrameRenderer.hpp
@@ -107,11 +107,11 @@ private:
     // ── Input dispatch ────────────────────────────────────────────────────────
     QPointF mapToAndroid(const QPointF& widgetPos) const;
 
-    void sendTouch(float ax, float ay, quint8 action, quint32 pointerId = 0);
+    void sendTouch(float ax, float ay, quint8 action, quint64 pointerId = 0);
     void sendScroll(float ax, float ay, float hScroll, float vScroll);
     void sendKeycode(quint32 keycode, quint32 action, quint32 metaState = 0);
 
-    QByteArray buildTouchMsg (float ax, float ay, quint8 action, quint32 id) const;
+    QByteArray buildTouchMsg (float ax, float ay, quint8 action, quint64 id) const;
     QByteArray buildScrollMsg(float ax, float ay, float hScroll, float vScroll) const;
     QByteArray buildKeyMsg   (quint32 keycode, quint32 action, quint32 meta) const;
